requirements.txt
anchor-topic==0.1.2
bigartm==0.9.2
dill==0.3.1.1
numpy==1.18.1
pandas==1.0.1
pytest==5.3.5
scikit-learn==0.22.1
scipy==1.4.1
topicnet==0.6.1
